chore: extract splitMessage from SummaryHandler to utility
Extracted `splitMessage` logic from `SummaryHandler` into a standalone utility function at `src/utils/message.util.ts`.
Also moved the related unit tests from `summary.handler.spec.ts` into a dedicated `message.util.spec.ts` file to properly separate concerns and maintain clean handler logic.
